Made the socket_store_t's RTT use a uint32_t instead of a uint16_t. This is because the timer.getNow() function returns a 32 bit integer. Also, got rid of the RTT pack, and now the RTT is sent from the client to the server inside the special ACK pack

+++ includes/socket.h
@@ -65,7 +65,7 @@ typedef struct socket_store_t{
     uint8_t lastRcvd;
     uint8_t nextExpected;
 
-    uint16_t RTT;
+    uint32_t RTT;
     uint16_t lastSentTime;
     uint8_t effectiveWindow;
 }socket_store_t;
